+++ tests/toisto/model/quiz/test_quiz.py
@@ -129,13 +129,18 @@ class QuizTest(QuizTestCase):
         quiz = self.create_quiz(self.concept, "fi", "fi", "Hän on;female", ["He ovat"], "pluralize")
         self.assertEqual("Give the [underline]plural[/underline] in Finnish", quiz.instruction())
 
-    def test_question_note_is_shown_when_question_language_equals_answer_language_but_quiz_type_is_answer(self):
-        """Test that a note is not shown when the question and answer languages are the same."""
-        quiz = self.create_quiz(self.concept, "fi", "fi", "Onko hän Bob?", ["On", "Ei"], "answer")
-        self.assertEqual("Answer the question in Finnish", quiz.instruction())
+    def test_question_note_is_shown_when_question_language_equals_answer_language_and_quiz_type_is_listen(self):
+        """Test that a note is shown when the question and answer languages are the same and the quiz type is listen."""
+        quiz = self.create_quiz(self.concept, "fi", "fi", "Suomi;country", ["Finland"], "listen")
+        self.assertEqual("Listen and write in Finnish (country)", quiz.instruction())
+
+    def test_question_note_is_shown_when_question_language_equals_answer_language_and_quiz_type_is_answer(self):
+        """Test that a note is shown when the question and answer languages are the same and the quiz type is answer."""
+        quiz = self.create_quiz(self.concept, "fi", "fi", "Onko hän Bob?;positive or negative", ["On", "Ei"], "answer")
+        self.assertEqual("Answer the question in Finnish (positive or negative)", quiz.instruction())
 
     def test_question_note_is_ignored_in_answer(self):
-        """Test that a note can be added to the question."""
+        """Test that a note in the answer is ignored."""
         quiz = self.create_quiz(self.concept, "en", "nl", "Jij bent", ["You are;singular"])
         self.assertEqual("You are", quiz.answer)
         self.assertEqual(("You are",), quiz.answers)