+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## 0.11.0 - [Unreleased]
 
+### Fixed
+
+- Fixed typo in the Dutch version of "to be": "zij zijn niet" and "ze zijn niet".
+
 ### Added
 
 - Added a books topic.