+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Fixed
 
-- Fixed typo in the Dutch version of "to be": "zij zijn niet" and "ze zijn niet".
+- Fixed a few typo's.
 
 ### Added